[develop] Refactoring to multiple source files completed
