+++ Makefile
@@ -1,22 +1,29 @@
-CC = gcc
-CC_FLAGS = -Wall -Werror -Wextra
-VPATH = src
-sources = game.c
-BUILD_PATH = build
-EXECUTABLE = $(BUILD_PATH)/snake
+CC := gcc
+CFLAGS := -Wall -Werror -Wextra
+VPATH := src
+SRC := $(wildcard src/*.c)
+OBJ := $(subst src/,build/,$(SRC:.c=.o))
+BUILD_PATH := build
+EXECUTABLE := $(BUILD_PATH)/snake
 
 .PHONY: all
 all: snake
 
+$(BUILD_PATH)/game.o: game.c
+	$(CC) -c $(CFLAGS) $< -o $@
+
+$(BUILD_PATH)/%.o: %.c %.h
+	$(CC) -c $(CFLAGS) $< -o $@
+
 .PHONY: snake
 snake: $(EXECUTABLE)
 
-$(EXECUTABLE): $(sources)
-	$(CC) $(CC_FLAGS) $^ -o $@
+$(EXECUTABLE): $(OBJ)
+	$(CC) $(CFLAGS) $^ -o $@
 
 .PHONY: clean
 clean:
-	rm -rf *.o $(EXECUTABLE)
+	rm -rf $(BUILD_PATH)/*.o $(EXECUTABLE)
 
 .PHONY: rebuild
 rebuild: clean snake
