+++ src/food.c
@@ -1,6 +1,7 @@
 #include "food.h"
 
 #include "util.h"
+#include "screen.h"
 
 Food GenerateFood(Snake* snake) {
     int x = (RandomUnsignedInt() << 1) % WIDTH;
