+++ src/food.h
@@ -1,18 +1,18 @@
 #ifndef FOOD_H
 #define FOOD_H
 
-#include "vector2d.h"
 #include "snake.h"
+#include "vector2d.h"
 
 #define FOOD_APPLE "🍎"
 
 struct Food {
     Vector pos;
-    char* symbol;
+    char *symbol;
 };
 
 typedef struct Food Food;
 
-Food GenerateFood(Snake* snake);
+Food GenerateFood(Snake *snake);
 
 #endif // FOOD_H
