+++ src/terminal.h
@@ -1,7 +1,20 @@
 #ifndef TERMINAL_H
 #define TERMINAL_H
 
+#include <poll.h>
+
+#define STDIN_FD 0
+#define STDOUT_FD 1
+#define STDERR_FD 2
+
+#define ESC 27
+
+#define COLOR_RED "\x1b[31m"
+#define COLOR_RESET "\x1b[0m"
+
 void EnableNonBlockingMode(int fd);
+void InitializeScreen(int fd);
+int ProcessInput(char *input, struct pollfd* readfds, int timeout);
 void AlternateScreenOn(int fd);
 void AlternateScreenOff(int fd);
 void RestoreTerminal(int fd);
