+++ src/util.h
@@ -1,8 +1,14 @@
+/*Some utility functions*/
 #ifndef UTIL_H
 #define UTIL_H
 
+/*Get random unsigned integer*/
 unsigned int RandomUnsignedInt(void);
+
+/*Custom function for finding non-negative remainder*/
 int Mod(int a, int b);
+
+/*Absolute value of the integer*/
 int Abs(int a);
 
 #endif // UTIL_H
